+++ exporters/otlp/test/otlp_grpc_log_exporter_test.cc
@@ -137,7 +137,8 @@ TEST_F(OtlpGrpcLogExporterTestPeer, ExportIntegrationTest)
                                                               '3', '2', '1', '0'};
   opentelemetry::trace::SpanId span_id{span_id_bin};
 
-  auto logger = provider->GetLogger("test");
+  const std::string schema_url{"https://opentelemetry.io/schemas/1.2.0"};
+  auto logger = provider->GetLogger("test", "", "opentelelemtry_library", "", schema_url);
   logger->Log(opentelemetry::logs::Severity::kInfo, "Log name", "Log message",
               {{"service.name", "unit_test_service"},
                {"tenant.id", "test_user"},